feat: UserController PostController

+++ app/Http/Controllers/Admin/PostController.php
@@ -8,6 +8,8 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
+
 
 class PostController extends Controller
 {
@@ -16,12 +18,19 @@ class PostController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+
+    public function index(Request $request)
     {
-        $posts = Post::orderBy('updated_at', 'DESC')->orderBy('created_at', 'DESC')->paginate(10);
+        $category_id = $request->query('category_id');
+
+        $query = Post::orderBy('updated_at', 'DESC')->orderBy('created_at', 'DESC');
+        $posts = $category_id ? $query->where('category_id', $category_id)->paginate(10) : $query->paginate(10);
         $categories = Category::all();
-        return view('admin.posts.index', compact('posts','categories'));
+        $selected_category = $category_id;
+        return view('admin.posts.index', compact('posts', 'categories', 'selected_category'));
     }
+
+
 
     /**
      * Show the form for creating a new resource.
@@ -64,6 +73,7 @@ class PostController extends Controller
         $post = new Post();
         $post->fill($data);
         $post->slug = Str::slug($post->title, '-');
+        $post->user_id = Auth::id();
         $post->save();
         return redirect()->route('admin.posts.show', $post)->with('message', "Post creato con successo")->with('type', "success");
     }
@@ -87,6 +97,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if ($post->user_id !== Auth::id()) {
+            return redirect()->route('admin.posts.index')->with('message', "Non sei autorizzato a modificare questo post")->with('type', "warning");
+        }
         $categories = Category::select('id', 'label')->get();
         return view('admin.posts.edit', compact('post', 'categories'));
     }
@@ -100,6 +113,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+
+
         $request->validate([
             'title' => ['required', 'string', 'min:5', 'max:50', Rule::unique('posts')->ignore($post->id)],
             'content' => 'required|string',
@@ -117,6 +132,8 @@ class PostController extends Controller
 
         $data = $request->all();
         $data['slug'] = Str::slug($data['title'], '-');
+
+        // if (array_key_exists('switch_author', $data)) $post->user_id = Auth::id();
         $post->update($data);
         return redirect()->route('admin.posts.index', $post)->with('message', "Post modificato con successo")->with('type', "success");
     }
@@ -129,6 +146,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if ($post->user_id !== Auth::id()) {
+            return redirect()->route('admin.posts.index')->with('message', "Non sei autorizzato ad eliminare questo post")->with('type', "warning");
+        }
         $post->delete();
         return redirect()->route('admin.posts.index')->with('message', 'Il post è stato eliminato correttamente')->with('type', 'success');
     }
